requirements.txt
annoying==0.1.0
appdirs==1.4.4
attr==0.3.2
attrs==22.2.0
av==10.0.0
azure_storage==0.37.0
bleach==5.0.1
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coverage==7.0.4
custom==0.1.1
dataloaders==0.0.1
debugpy==1.6.3
deepdiff==6.2.3
defusedxml==0.7.1
diskcache==5.4.0
dj_rest_auth==2.2.6
Django==3.2.16
django_annoying==0.10.6
django_auth_ldap==4.1.0
django_cors_headers==3.6.0
django_debug_toolbar==3.2.1
django_filter==2.4.0
django_ranged_fileresponse==0.1.2
django_rq==2.5.1
django_sendfile==0.3.11
django_storages==1.13.2
djangorestframework==3.14.0
dnspython==2.2.1
drf_dynamic_fields==0.3.0
drf_flex_fields==0.9.5
drf_spectacular==0.25.1
drf_yasg==1.20.0
fakeredis==2.4.0
fiftyone==0.18.0
fiftyone_db==0.4.0
flaml==1.1.1
flash==1.0.3
furl==2.1.3
health_check==3.4.1
htmlmin==0.1.12
inflection==0.5.1
json2html==1.3.0
jsonpickle==3.0.1
jsonschema==3.2.0
keras==2.11.0
keys==0.1-test
label_studio_converter==0.0.48rc0
label_studio_ml==1.0.9
label_studio_tools==0.0.1
launchdarkly_server_sdk==7.5.0
ldclient==0.0.1
locustio==0.999
lxml==4.9.2
mock==5.0.1
moto==4.1.0
mrcnn==0.2
natsort==8.2.0
networks==0.3.7
numpy==1.23.3
opencv_python_headless==4.7.0.68
packaging==21.3
pandas==1.3.5
pdf2image==1.16.2
Pillow==9.4.0
protobuf==4.21.12
pycocotools==2.0.6
pydantic==1.8.2
pydicom==2.3.1
pytest==7.2.0
pytest_env==0.8.1
python-dotenv==0.21.0
python_dateutil==2.8.2
python_json_logger==2.0.4
pytz==2019.3
pyunpack==0.3
PyYAML==6.0
requests==2.27.1
requests_mock==1.10.0
rq==1.10.1
ruamel.base==1.0.0
scikit_image==0.19.3
scikit_learn==1.2.0
scipy==1.10.0
sentry_sdk==1.12.1
setuptools==61.2.0
Shapely==2.0.0
silk==0.1
skimage==0.0
SQLAlchemy==1.4.46
toml==0.10.2
tools==0.1.9
torch==1.13.1
torchvision==0.14.1
tqdm==4.63.0
tusclient==0.2.1
typer==0.7.0
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
wandb==0.13.7
xmljson==0.2.0
